changing code list config file structure
